requirements.txt
langchain
python-dotenv==1.0.0
streamlit
langchain_openai
langchain_community
plotly
sqlalchemy
